refactor(backend): split media repository row mappers
